requirements.txt
alembic==1.8.0
aniso8601==9.0.1
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask-Caching==2.0.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask==2.1.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
iniherit==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow-sqlalchemy==0.28.0
marshmallow==3.17.0
nose==1.3.7
packaging==21.3
pip==22.1.2
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.1
setuptools==60.2.0
six==1.16.0
SQLAlchemy==1.4.39
tabulate==0.8.2
text-unidecode==1.3
urllib3==1.26.10
Werkzeug==2.1.2
wheel==0.37.1
zipp==3.8.0
